requirements.txt
langchain
langchain-community
langgraph
langserve
python-dotenv
langchain_google_genai
fastapi
uvicorn